mavgen_wlua: improved truncation handling
The previous truncation handling passed too short buffers for TvbRange:float()
and :string(), leading to incorrect results. This change properly zero-pads
the payload before parsing (if necessary).

This change also separates the field display type from the value type. This
allows for parsing a field as a float, but displaying it as an integer (which
is required to show enum values); at least command parameters have enums sent
as floats.